Update flap_nstx_gpi.py
Added the option of adding the normalized poloidal flux as a coordinate. It is interpolated first to the image of the GPI then it is upsampled (interpolated in time) to the time resolution of the GPI image.
Interpolation along the time vector with slicing was slow, I used np.interp() instead.
scipy.interpolate.griddata was used for spatial interpolation and its cubic method was utilized.