Update flap_nstx_gpi.py
Added the options of 'Device theta' and finalized the 'Flux theta' options:

Device theta:
Geometrical poloidal angle coordinate calculated from the magnetic axis and the geometry of the GPI measurement.

Flux theta:
Poloidal angle coordinate calculated solely from the magnetic geometry. Since the angle is not defined outside the separatrix and diverges at the separatrix, an extrapolation had to be used. The angles are taken from the position of the 0.9*a (minor radius) and the geometry is zoom out as the separatrix is shaped. This start from 0.9a --> 2.0a. In this way the GPI geometry stays within the range all the time.
The beginning and the end of the shot is still problematic, this approximation is not done there. There, wrong coordinates are returned, but GPI is only measured during the time when there is a gas puff, which is not at the start or end of the plasma.