Major modification
Importing modification for ease of use (not as easy as expected, but better than before). If only the flap_nstx is modified, the flap storage doesn't have to be reloaded. a del flap_nstx is needed before
Added the multidim_detrend in gpi_tools
Added the filament finder in the velocity_analysis