Mostly analysis and bugfix related modifications, nothing substantial
tools: Bugfixes in the structure finder, added the signal_windowed_avg_err function, which can return a smoothed signal with error bars calculated in a window.
velocity_analysis: mostly bugfixes and reorganizing the code. Added the possibility of plotting a time range within a larger time range from a file. Also the filenames are 6digits in time now instead of almost 15.
video: bugfix related to caching the data. The exp_id is also given to the flap methods in order to handle the FLAP storage when multiple events are stored in there.
Added the plot_nstx_gpi_velocity_results.py function