Multiple code modifications
flap_nstx_gpi.py: The spatial calibration is changed to the one perpendicular to the magnetic coordinate.

calculate_nstx_gpi_frame_by_frame_velocity.py: changed the spatial coordinates to be consistent with the data

claculate_sde_velocity.py: A more generalized version of the gpi_frame_by_frame velocity calculation method. It takes a flap dataobject or a name of the dataobject, determines the spatial calibration from the spatial coordinates and returns a dict with the calculated parameters.

analyze_hesel_data.py: Modified for the new SDE velocity calculation.